Restructure BE

+++ src/main/java/com/example/onlyfanshop_be/repository/ProductRepository.java
@@ -1,11 +1,10 @@
 package com.example.onlyfanshop_be.repository;
 
-
 import com.example.onlyfanshop_be.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
